+++ roman_discovery/__init__.py
@@ -1,5 +1,3 @@
-"""Top-level package for Roman Discovery."""
-
 __author__ = """Roman Imankulov"""
 __email__ = "roman.imankulov@gmail.com"
 __version__ = "0.3.1"