+++ roman_discovery/__init__.py
@@ -1,6 +1,6 @@
 __author__ = """Roman Imankulov"""
 __email__ = "roman.imankulov@gmail.com"
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 
 
 from roman_discovery.discovery import IRule, ModuleRule, ObjectRule, discover